requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
markdownify>=0.11.0
Pillow>=9.0.0
PyYAML>=6.0.0
markdown>=3.3.0 